requirements.txt
bleak==0.20.2
pydBeacon==1.0.0